// apps/frontend/src/api/generated/model/problem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Avito Recap API
 * API MVP «Итоги года».
 *
 * OpenAPI spec version: 0.1.0
 */

export interface Problem {
  type: string;
  title: string;
  status: number;
  code: string;
  detail?: string;
  instance?: string;
  requestId?: string;
}